requirements.txt
flask==2.1.1
flask-restx==0.5.1
black==22.12.0
werkzeug==2.1.2
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.9.3
pytest==7.2.0
gunicorn==20.1.0
pytest-cov==3.0.0
flake8==4.0.1
isort==5.10.1
pytest-xdist==2.5.0